Refactor part 1
Summary:

Test Plan:
