+++ packages/app/constants/deploymentEnvs.ts
@@ -1,0 +1,5 @@
+export enum DeploymentEnvs {
+  DEVELOPMENT = "DEVELOPMENT",
+  STAGING = "STAGING",
+  PRODUCTION = "PRODUCTION",
+}
