+++ packages/web/web.tsx
@@ -1,56 +1,22 @@
 #! /usr/bin/env -S deno run --allow-net=localhost:8000 --allow-env
 
-import React from "react";
-import { renderToReadableStream } from "react-dom/server";
-import { appRoutes, isographAppRoutes } from "packages/app/routes.ts";
-import { ServerRenderedPage } from "packages/app/server/components/ServerRenderedPage.tsx";
-import { AppRoot } from "packages/app/AppRoot.tsx";
 import { serveDir } from "@std/http";
-import type { ServerProps } from "packages/app/contexts/AppEnvironmentContext.tsx";
-import { ClientRoot } from "packages/app/ClientRoot.tsx";
+import { appRoutes, isographAppRoutes } from "packages/app/routes.ts";
 import { graphQLHandler } from "packages/graphql/graphqlServer.ts";
-import { getIsographEnvironment } from "packages/app/server/isographEnvironment.ts";
-
 import { getLogger } from "packages/logger.ts";
-import { AppEnvironmentProvider } from "packages/app/contexts/AppEnvironmentContext.tsx";
-import { useLazyReference, useResult } from "@isograph/react";
 import { matchRouteWithParams } from "packages/app/contexts/RouterContext.tsx";
-import { AssemblyAI } from "assemblyai";
 import { getConfigurationVariable } from "packages/getConfigurationVariable.ts";
 import { BfCurrentViewer } from "packages/bfDb/classes/BfCurrentViewer.ts";
-import type { BfIsographEntrypoint } from "lib/BfIsographEntrypoint.ts"; // Import added here
+import { initializeContentCollections } from "packages/web/initializeContent.ts";
+import { DeploymentEnvs } from "packages/app/constants/deploymentEnvs.ts";
+import {
+  handleAppRoute,
+  handleAssemblyAI,
+  handleIsographRoute,
+  handleLogout,
+} from "packages/web/handlers/routeHandlers.ts";
 
 const logger = getLogger(import.meta);
-
-function IsographHeaderComponent(
-  { entrypoint }: {
-    entrypoint: BfIsographEntrypoint;
-  },
-) {
-  const { fragmentReference } = useLazyReference(entrypoint, {});
-  const result = useResult(fragmentReference);
-  const title = result.title;
-  return <title className="dynamic">{title}</title>;
-}
-
-function getIsographHeaderComponent(
-  environment: ServerProps,
-  entrypoint: BfIsographEntrypoint,
-) {
-  return function IsographHeader() {
-    return (
-      <AppEnvironmentProvider {...environment}>
-        <IsographHeaderComponent entrypoint={entrypoint} />
-      </AppEnvironmentProvider>
-    );
-  };
-}
-
-export enum DeploymentEnvs {
-  DEVELOPMENT = "DEVELOPMENT",
-  STAGING = "STAGING",
-  PRODUCTION = "PRODUCTION",
-}
 
 export type Handler = (
   request: Request,
@@ -64,58 +30,11 @@ if (getConfigurationVariable("BF_ENV") !== DeploymentEnvs.DEVELOPMENT) {
   appRoutes.delete("/ui");
 }
 
-const configurationVariableKeys = [
-  "POSTHOG_API_KEY",
-];
-
-const configurationVariables = configurationVariableKeys.reduce((acc, key) => {
-  acc[key] = getConfigurationVariable(key);
-  return acc;
-}, {} as Record<string, string | undefined>);
-
 // Register each app route in the routes Map
 for (const entry of appRoutes.entries()) {
-  const [path, { Component: { HeaderComponent: RouteHeaderComponent } }] =
-    entry;
+  const [path] = entry;
   routes.set(path, async function AppRoute(request, routeParams) {
-    const reqUrl = new URL(request.url);
-    const initialPath = reqUrl.pathname;
-    const queryParams = Object.fromEntries(reqUrl.searchParams.entries());
-    const isographServerEnvironment = await getIsographEnvironment(request);
-
-    const clientEnvironment = {
-      initialPath,
-      queryParams,
-      routeParams,
-      path,
-      ...configurationVariables,
-    };
-
-    const serverEnvironment: ServerProps = {
-      ...clientEnvironment,
-      IS_SERVER_RENDERING: true,
-      isographServerEnvironment,
-    };
-
-    const HeaderComponent = RouteHeaderComponent ?? AppRoot.HeaderComponent;
-    const headerElement = React.createElement(HeaderComponent);
-
-    const appElement = React.createElement(
-      ClientRoot,
-      serverEnvironment,
-      React.createElement(AppRoot),
-    );
-
-    const element = React.createElement(
-      ServerRenderedPage,
-      { headerElement, environment: clientEnvironment },
-      appElement,
-    );
-
-    const stream = await renderToReadableStream(element);
-    return new Response(stream, {
-      headers: { "content-type": "text/html; charset=utf-8" },
-    });
+    return await handleAppRoute(request, routeParams, path);
   });
 }
 
@@ -123,54 +42,7 @@ for (const entry of appRoutes.entries()) {
 for (const [path, entrypoint] of isographAppRoutes.entries()) {
   logger.debug(`Registering ${path}`, entrypoint);
   routes.set(path, async function AppRoute(request, routeParams) {
-    const reqUrl = new URL(request.url);
-    const initialPath = reqUrl.pathname;
-    const queryParams = Object.fromEntries(reqUrl.searchParams.entries());
-    const isographServerEnvironment = await getIsographEnvironment(request);
-    using cv = BfCurrentViewer.createFromRequest(import.meta, request);
-    logger.debug("path", path);
-    logger.debug("entrypoint", entrypoint);
-    const ph = await cv.getPosthogClient();
-    const featureFlags = await ph.backendClient?.getAllFlags(cv.bfGid);
-
-    const clientEnvironment = {
-      initialPath,
-      queryParams,
-      routeParams,
-      path,
-      featureFlags,
-      ...configurationVariables,
-    };
-
-    const serverEnvironment: ServerProps = {
-      ...clientEnvironment,
-      IS_SERVER_RENDERING: true,
-      isographServerEnvironment,
-    };
-
-    // Because this route is isograph-based, we dynamically generate a header component
-    const HeaderComponent = getIsographHeaderComponent(
-      serverEnvironment,
-      entrypoint,
-    );
-    const headerElement = React.createElement(HeaderComponent);
-
-    const appElement = React.createElement(
-      ClientRoot,
-      serverEnvironment,
-      React.createElement(AppRoot),
-    );
-
-    const element = React.createElement(
-      ServerRenderedPage,
-      { headerElement, environment: clientEnvironment },
-      appElement,
-    );
-
-    const stream = await renderToReadableStream(element);
-    return new Response(stream, {
-      headers: { "content-type": "text/html; charset=utf-8" },
-    });
+    return await handleIsographRoute(request, routeParams, path, entrypoint);
   });
 }
 
@@ -187,50 +59,11 @@ routes.set("/static/:filename+", function staticHandler(req) {
 // GraphQL handler
 routes.set("/graphql", graphQLHandler);
 
-// Example for handling form uploads with AssemblyAI
-routes.set("/assemblyai", async (req) => {
-  try {
-    const formData = await req.formData();
-    const file = formData.get("file") as File;
-
-    if (!file) {
-      return new Response("No file provided", { status: 400 });
-    }
-
-    const client = new AssemblyAI({
-      apiKey: getConfigurationVariable("ASSEMBLY_AI_KEY") as string,
-    });
-
-    const data = { audio: file };
-    const transcript = await client.transcripts.transcribe(data);
-    const words = transcript.words;
-
-    return new Response(JSON.stringify(words), {
-      headers: { "Content-Type": "application/json" },
-    });
-  } catch (error) {
-    logger.error(error);
-    return new Response("Internal server error", { status: 500 });
-  }
-});
+// AssemblyAI handler
+routes.set("/assemblyai", handleAssemblyAI);
 
 // Simple logout route that clears cookies
-routes.set("/logout", function logoutHandler() {
-  const headers = new Headers();
-  headers.set("location", "/");
-  headers.set(
-    "set-cookie",
-    "bfgat=; path=/; expires=Thu, 01 Jan 1970 00:00:00 GMT",
-  );
-  headers.set(
-    "set-cookie",
-    "bfgrt=; path=/; expires=Thu, 01 Jan 1970 00:00:00 GMT",
-  );
-  return new Response(null, {
-    status: 302,
-    headers,
-  });
-});
+routes.set("/logout", handleLogout);
 
 // Fallback default route
 function defaultRoute() {
@@ -253,9 +86,6 @@ function matchRoute(
 }
 
 // Initialize content collections
-import { initializeContentCollections } from "packages/web/initializeContent.ts";
-
-// Create viewer for content initialization
 const contentInitCv = BfCurrentViewer
   .__DANGEROUS_USE_IN_SCRIPTS_ONLY__createOmni(import.meta);
 
@@ -272,9 +102,6 @@ function staticHandler(req: Request) {
   });
 }
 
-// Use the port from environment or default 8000
-const port = Number(Deno.env.get("WEB_PORT") ?? 8000);
-
 async function handleDomains(req: Request) {
   const reqUrl = new URL(req.url);
   const domain = Deno.env.get("SERVE_PROJECT") ?? reqUrl.hostname;
@@ -290,6 +117,9 @@ async function handleDomains(req: Request) {
   }
   return null;
 }
+
+// Use the port from environment or default 8000
+const port = Number(Deno.env.get("WEB_PORT") ?? 8000);
 
 if (import.meta.main) {
   Deno.serve({ port }, async (req) => {
